Set __request__ on objects w/ SQLA ORM loader
In several cases, resource instances end up floating around that were
fetched straight from the ORM and weren’t produced by some other
resource’s __getattr__, so the __request__ attribute was not set on the
instance, but the instance uses it for something. To prevent this, use
Pyramid’s get_current_request() to set it on load. This is an ugly hack
as Pyramid strongly discourages use of this function, but until we can
come up with a better method, it will have to do