Fix for when no dependencies defined
The configuration parser calls the eval_with_deps function from
tools.py, which will either use a dictionary of dependencies passed in
as an argument or will fall back to calling dependencies() from
config.py to use the dependencies defined in api.yaml. Since
config.load_api_config(…) has not been called at the time
eval_with_deps is called, in parser.py we supply the list of
dependencies to the eval_with_deps call so that we circumvent calling
config.dependencies() (which will clearly result in a KeyError if
_api_config is {} because it has not been configured yet — see prior
definition of config.dependencies())

HOWEVER — if no dependencies are declared in the api config, then the
parser will call eval_with_deps with dependencies={}. Then “if not
dependencies” (tools.py line 32) will evaluate to True and we end up
calling config.dependencies() which goes catastrophically because it
has not been configured yet. Rekt.

So, in config.py, return {} to avoid a KeyError if there is no
“dependencies” in the API config (and return “empty” values for the
other things as well). Also warn about sofa not having been configured
if sqla_session() or root_collections() are called.

Bump to v0.5