Only check access token once per request
To avoid dumb scenarios where sofa checks the access token several
thousand times in one request (e.g. if we are listing a lot of objects,
it would verify the access token for every single attribute on every
single object), set a flag on the request. This speeds up those kinds
of scenarios by quite a bit and reduces unnecessary database traffic