Pass APIAttr object to validator validate() as argument
In addition to passing APIValidator.validate() a value to validate,
also pass in the APIAttr object that the validator is doing the
validation for. That way, the validator can directly access
APIAttr.key, removing the necessity for roundabout ways of setting
APIValidator.attr_name.  Additionally, passing the APIAttr in when
validate() is called (as opposed to on __init__) removes the potential
for bugs where the same validator object is used for multiple
attributes (in the previous setup, each attribute would try to set
`attr_name` on that validator object with its own name, and overwrite
the other validators).