Remove default `active` attribute
Trying to make resources quasi-deletable by default causes lots of
issues and probably wasn’t a great design decision. Users of sofa
should consciously decide that they want to be able to un-delete things
and have `active` attributes.

A really big issue is that although sofa accounts for `active` and
won’t return inactive resources in the API, we don’t have a default
filter on database queries that prevents inactive resources from
showing up if someone is using SQLAlchemy in the Python app to access
the database directly. This is very bad. Sofa might say a user doesn’t
exist, but technically the ID is still in the database, so a user might
be able to get an auth token or do be associated with something since
the user is technically still there, even though the sofa user thinks
it has been deleted. We could introduce default filters with
SQLAlchemy, but it will introduce more complexity than is worth, and…

an even bigger issue is that users really should know, very clearly,
that data isn’t being deleted. Sometimes you don’t want to be able to
un-delete things, and currently, there is no way to actually delete
things with sofa. Users should be able to decide whether or not they
want “phantom deletes.”

This is a breaking change :( but better sooner than later.