Add support for default filters
If a sofa user chooses to implement an `active` field, he/she can
specify a default filter in the api.yaml file so that by default, only
active resources will be returned in a query. Note however that default
filters will only apply to list requests, and if an API caller uses a
specific resource ID (e.g. someone tries to update a resource with
active==0), the default filter will not be applied (and hence the
collection won’t 404 if trying to access a resource that doesn’t match
the default filter)